feat: implement discover test feature with filtering, searching, and test card display
