+++ components/container/select-module-type.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { memo } from "react";
 import { ModuleType } from "@/types/class";
 import {
   Select,
@@ -10,33 +11,41 @@ import {
 } from "@/components/ui/select";
 import { cn } from "@/lib/utils";
 
-const SelectModuleType = ({
-  value,
-  onChange,
-  placeholder = "Select module type",
-  className,
-}: {
-  value: ModuleType;
-  onChange: (val: ModuleType) => void;
-  placeholder?: string;
-  className?: string;
-}) => {
-  return (
-    <Select value={value} onValueChange={(val) => onChange(val as ModuleType)}>
-      <SelectTrigger
-        className={cn("h-11 w-full border-[#FFFFFF66] bg-[#333333]", className)}
+export const SelectModuleType = memo(
+  ({
+    value,
+    onChange,
+    placeholder = "Select module type",
+    className,
+  }: {
+    value: ModuleType;
+    onChange: (val: ModuleType) => void;
+    placeholder?: string;
+    className?: string;
+  }) => {
+    return (
+      <Select
+        value={value}
+        onValueChange={(val) => onChange(val as ModuleType)}
       >
-        <SelectValue placeholder={placeholder} />
-      </SelectTrigger>
-      <SelectContent className="border-[#FFFFFF66] p-1">
-        {Object.values(ModuleType).map((type) => (
-          <SelectItem className="rounded-xl" key={type} value={type}>
-            {type.charAt(0).toUpperCase() + type.slice(1)}
-          </SelectItem>
-        ))}
-      </SelectContent>
-    </Select>
-  );
-};
+        <SelectTrigger
+          className={cn(
+            "h-11 w-full border-[#FFFFFF66] bg-[#333333]",
+            className,
+          )}
+        >
+          <SelectValue placeholder={placeholder} />
+        </SelectTrigger>
+        <SelectContent className="border-[#FFFFFF66] p-1">
+          {Object.values(ModuleType).map((type) => (
+            <SelectItem className="rounded-xl" key={type} value={type}>
+              {type.charAt(0).toUpperCase() + type.slice(1)}
+            </SelectItem>
+          ))}
+        </SelectContent>
+      </Select>
+    );
+  },
+);
 
-export default SelectModuleType;
+SelectModuleType.displayName = "SelectModuleType";
