+++ types/discover-test.ts
@@ -1,0 +1,6 @@
+export type Test = {
+  id: string;
+  image?: string | null;
+  name: string;
+  attempts: number;
+};
